+++ agent/src/main/java/com/teamscale/jacoco/agent/options/ArtifactoryConfig.java
@@ -1,12 +1,9 @@
-package com.teamscale.jacoco.agent.upload.artifactory;
+package com.teamscale.jacoco.agent.options;
 
 import com.teamscale.client.CommitDescriptor;
 import com.teamscale.client.StringUtils;
 import com.teamscale.jacoco.agent.commit_resolution.git_properties.GitPropertiesLocator;
 import com.teamscale.jacoco.agent.commit_resolution.git_properties.InvalidGitPropertiesException;
-import com.teamscale.jacoco.agent.options.AgentOptionParseException;
-import com.teamscale.jacoco.agent.options.AgentOptionsParser;
-import com.teamscale.jacoco.agent.options.FilePatternResolver;
 import okhttp3.HttpUrl;
 import org.conqat.lib.commons.collections.Pair;
 
@@ -48,10 +45,7 @@ public class ArtifactoryConfig {
 			throws AgentOptionParseException {
 		switch (key) {
 			case "artifactory-url":
-				options.url = AgentOptionsParser.parseUrl(value);
-				if (options.url == null) {
-					throw new AgentOptionParseException("Invalid URL given for option 'artifactory-url'");
-				}
+				options.url = AgentOptionsParser.parseUrl(key, value);
 				return true;
 			case "artifactory-user":
 				options.user = value;