+++ agent/src/test/java/com/teamscale/jacoco/agent/options/AgentOptionsTest.java
@@ -4,7 +4,6 @@ import com.teamscale.client.CommitDescriptor;
 import com.teamscale.client.TeamscaleServer;
 import com.teamscale.jacoco.agent.util.TestUtils;
 import com.teamscale.report.util.CommandLineLogger;
-
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -17,8 +16,8 @@ import java.util.function.Predicate;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /** Tests the {@link AgentOptions}. */
 public class AgentOptionsTest {
@@ -114,11 +113,11 @@ public class AgentOptionsTest {
 	@Test
 	public void testHttpServerOptionsWithCoverageViaHttp() throws AgentOptionParseException {
 		AgentOptions agentOptions = getAgentOptionsParserWithDummyLogger().parse("mode=TESTWISE,class-dir=.," +
-				"http-server-port=8081,coverage-via-http=true");
+				"http-server-port=8081,tia-mode=http");
 		assertThat(agentOptions.getHttpServerPort()).isEqualTo(8081);
-		assertThat(agentOptions.getTestWiseCoverageMode()).isEqualTo(ETestWiseCoverageMode.HTTP);
+		assertThat(agentOptions.getTestwiseCoverageMode()).isEqualTo(ETestwiseCoverageMode.HTTP);
 	}
-	
+
 	/** Tests setting ignore-uncovered-classes works. */
 	@Test
 	public void testIgnoreUncoveredClasses() throws AgentOptionParseException {
@@ -132,7 +131,14 @@ public class AgentOptionsTest {
 		AgentOptions agentOptions = getAgentOptionsParserWithDummyLogger().parse("");
 		assertFalse(agentOptions.shouldIgnoreUncoveredClasses());
 	}
-	
+
+	/** Tests default for ignore-uncovered-classes is false. */
+	@Test
+	public void shouldAllowMinusForEnumConstants() throws AgentOptionParseException {
+		AgentOptions agentOptions = getAgentOptionsParserWithDummyLogger().parse("tia-mode=exec-file");
+		assertThat(agentOptions.getTestwiseCoverageMode()).isEqualTo(ETestwiseCoverageMode.EXEC_FILE);
+	}
+
 	/** Tests that supplying both revision and commit info is forbidden. */
 	@Test
 	public void testBothRevisionAndCommitSupplied() throws URISyntaxException {