+++ system-tests/tia-client-test/src/main/java/testframework/CustomTestFramework.java
@@ -7,6 +7,7 @@ import com.teamscale.report.testwise.model.ETestExecutionResult;
 import com.teamscale.tia.client.AgentHttpRequestFailedException;
 import com.teamscale.tia.client.RunningTest;
 import com.teamscale.tia.client.TestRun;
+import com.teamscale.tia.client.TestRunWithClusteredSuggestions;
 import com.teamscale.tia.client.TiaAgent;
 import okhttp3.HttpUrl;
 import systemundertest.SystemUnderTest;
@@ -39,19 +40,19 @@ public class CustomTestFramework {
 	/** Talks to the TIA agent and runs all impacted tests. Also uploads a coverage report at the end. */
 	public void runTestsWithTia() throws AgentHttpRequestFailedException {
 		TiaAgent agent = new TiaAgent(false, HttpUrl.get("http://localhost:" + agentPort));
-		TestRun testRun = agent.startTestRun(
-				allTests.keySet().stream().map(name -> new ClusteredTestDetails(name, name, null, null)).collect(
-						toList()));
+		TestRunWithClusteredSuggestions testRun = agent.startTestRun(
+				allTests.keySet().stream().map(name -> new ClusteredTestDetails(name, name, null, null))
+						.collect(toList()));
 
-		for (PrioritizableTestCluster cluster : testRun.getPrioritizedTests()) {
+		for (PrioritizableTestCluster cluster : testRun.getPrioritizedClusters()) {
 			for (PrioritizableTest test : cluster.tests) {
 				Runnable runnable = allTests.get(test.uniformPath);
 				RunningTest runningTest = testRun.startTest(test.uniformPath);
 				try {
 					runnable.run();
-					runningTest.endTestNormally(new TestRun.TestResultWithMessage(ETestExecutionResult.PASSED, ""));
+					runningTest.endTest(new TestRun.TestResultWithMessage(ETestExecutionResult.PASSED, ""));
 				} catch (Throwable t) {
-					runningTest.endTestNormally(
+					runningTest.endTest(
 							new TestRun.TestResultWithMessage(ETestExecutionResult.FAILURE, t.getMessage()));
 				}
 			}