+++ tia-client/src/main/java/com/teamscale/tia.client/TestRun.java
@@ -1,34 +1,20 @@
 package com.teamscale.tia.client;
 
-import com.teamscale.client.PrioritizableTestCluster;
 import com.teamscale.report.testwise.model.ETestExecutionResult;
 
-import java.util.List;
-
 /**
- * Represents a run of prioritized and selected tests as reported by the TIA. Use this class to report test start and
- * end events and upload test-wise coverage to Teamscale.
+ * Use this class to report test start and end events and upload testwise coverage to Teamscale.
  * <p>
- * The caller of this class should retrieve the tests to execute from {@link #getPrioritizedTests()}, run them (in the
- * given order if possible) and report test start and end events via {@link #startTest(String)} )}.
- * <p>
- * After having run all tests, call {@link #endTestRun()} to create a test-wise coverage report and upload it to
- * Teamscale.
+ * After having run all tests, call {@link #endTestRun()} to create a testwise coverage report and upload it to
+ * Teamscale. This requires that you configured the agent to upload coverage to Teamscale
+ * (`tia-mode=teamscale-upload`).
  */
-@SuppressWarnings("unused")
 public class TestRun {
 
 	private final ITestwiseCoverageAgentApi api;
-	private final List<PrioritizableTestCluster> prioritizedTests;
 
-	TestRun(ITestwiseCoverageAgentApi api,
-			List<PrioritizableTestCluster> prioritizedTests) {
+	TestRun(ITestwiseCoverageAgentApi api) {
 		this.api = api;
-		this.prioritizedTests = prioritizedTests;
-	}
-
-	public List<PrioritizableTestCluster> getPrioritizedTests() {
-		return prioritizedTests;
 	}
 
 	/**
@@ -49,12 +35,12 @@ public class TestRun {
 	}
 
 	/**
-	 * Informs the test-wise coverage agent that a new test is about to start.
+	 * Informs the testwise coverage agent that a new test is about to start.
 	 *
 	 * @throws AgentHttpRequestFailedException if communicating with the agent fails or in case of internal errors. In
 	 *                                         this case, the agent probably doesn't know that this test case was
 	 *                                         started, so its coverage is lost. This method already retries the request
-	 *                                         once, so this is likely a terminal failure. The caller should record this
+	 *                                         once, so this is likely a terminal failure. The caller should log this
 	 *                                         problem appropriately. Coverage for subsequent test cases could, however,
 	 *                                         potentially still be recorded. Thus, the caller should continue with test
 	 *                                         execution and continue informing the coverage agent about further test
@@ -67,13 +53,14 @@ public class TestRun {
 	}
 
 	/**
-	 * Informs the test-wise coverage agent that the caller has finished running all tests and instructs it to upload
-	 * the recorded test-wise coverage to Teamscale.
+	 * Informs the testwise coverage agent that the caller has finished running all tests and should upload coverage to
+	 * Teamscale. Only call this if you configured the agent to upload coverage to Teamscale
+	 * (`tia-mode=teamscale-upload`). Otherwise, this method will throw an exception.
 	 *
 	 * @throws AgentHttpRequestFailedException if communicating with the agent fails or in case of internal errors. This
 	 *                                         method already retries the request once, so this is likely a terminal
-	 *                                         failure. The recorded coverage is likely lost. The caller should record
-	 *                                         this problem appropriately.
+	 *                                         failure. The recorded coverage is likely lost. The caller should log this
+	 *                                         problem appropriately.
 	 */
 	public void endTestRun() throws AgentHttpRequestFailedException {
 		AgentCommunicationUtils.handleRequestError(api::testRunFinished,