+++ tia-client/src/main/java/com/teamscale/tia.client/TestWiseCoverageRunListener.java
@@ -35,7 +35,7 @@ public class TestWiseCoverageRunListener extends RunListener {
 		}
 
 		TiaAgent agent = new TiaAgent(false, HttpUrl.get(agentUrl));
-		testRun = agent.startTestRun();
+		testRun = agent.startTestRunWithoutTestSelection();
 	}
 
 	@FunctionalInterface
@@ -51,9 +51,9 @@ public class TestWiseCoverageRunListener extends RunListener {
 	private void handleErrors(Action action) {
 		try {
 			action.run();
-		} catch (Throwable t) {
+		} catch (Exception e) {
 			System.err.println("Encountered an error while recording test-wise coverage:");
-			t.printStackTrace(System.err);
+			e.printStackTrace(System.err);
 		}
 	}
 
@@ -74,7 +74,7 @@ public class TestWiseCoverageRunListener extends RunListener {
 	public void testFinished(Description description) {
 		handleErrors(() -> {
 			if (runningTest != null) {
-				runningTest.endTestNormally(new TestRun.TestResultWithMessage(ETestExecutionResult.PASSED, null));
+				runningTest.endTest(new TestRun.TestResultWithMessage(ETestExecutionResult.PASSED, null));
 				runningTest = null;
 			}
 		});
@@ -84,7 +84,7 @@ public class TestWiseCoverageRunListener extends RunListener {
 	public void testFailure(Failure failure) {
 		handleErrors(() -> {
 			if (runningTest != null) {
-				runningTest.endTestNormally(
+				runningTest.endTest(
 						new TestRun.TestResultWithMessage(ETestExecutionResult.FAILURE, failure.getTrace()));
 				runningTest = null;
 			}
@@ -95,7 +95,7 @@ public class TestWiseCoverageRunListener extends RunListener {
 	public void testAssumptionFailure(Failure failure) {
 		handleErrors(() -> {
 			if (runningTest != null) {
-				runningTest.endTestNormally(
+				runningTest.endTest(
 						new TestRun.TestResultWithMessage(ETestExecutionResult.FAILURE, failure.getTrace()));
 				runningTest = null;
 			}